Extract main test to a separate file
