+++ api/main_test.go
@@ -1,0 +1,32 @@
+package api_test
+
+import (
+	"flag"
+	"log"
+	"os"
+	"testing"
+
+	"github.com/ory/dockertest/v3"
+	"github.com/treeverse/lakefs/testutil"
+)
+
+var (
+	pool        *dockertest.Pool
+	databaseURI string
+)
+
+func TestMain(m *testing.M) {
+	flag.Parse()
+
+	var err error
+	pool, err = dockertest.NewPool("")
+	if err != nil {
+		log.Fatalf("Could not connect to Docker: %s", err)
+	}
+
+	var closer func()
+	databaseURI, closer = testutil.GetDBInstance(pool)
+	code := m.Run()
+	closer() // cleanup
+	os.Exit(code)
+}
